Refactor `SignTransaction` type-specific UI blocks
Creates a component `SignTransactionInfoProvider`,
which exposes to its children some UI blocks,
specific to the current type of transaction.
Allows different signing flows (keyrings/Ledger/others in the future)
to show these UI blocks when appropriate,
without invalidating any transaction type-specific UI state.
