+++ ui/pages/SignTransaction.tsx
@@ -9,33 +9,17 @@ import {
   signTransaction,
 } from "@tallyho/tally-background/redux-slices/transaction-construction"
 import { getAccountTotal } from "@tallyho/tally-background/redux-slices/selectors"
-import { parseERC20Tx } from "@tallyho/tally-background/lib/erc20"
-import SignTransactionSwapAssetBlock from "../components/SignTransaction/SignTransactionSwapAssetBlock"
-import SignTransactionSignBlock from "../components/SignTransaction/SignTransactionSignBlock"
 import {
   useBackgroundDispatch,
   useBackgroundSelector,
   useAreKeyringsUnlocked,
 } from "../hooks"
-import SignTransactionTransferBlock from "../components/SignTransaction/SignTransactionTransferBlock"
 import SignTransactionContainer from "../components/SignTransaction/SignTransactionContainer"
-import SignTransactionApproveSpendAssetBlock from "../components/SignTransaction/SignTransactionApproveSpendAssetBlock"
-import SignTransactionTransferDetail from "../components/SignTransaction/SignTransactionTransferDetail"
+import SignTransactionInfoProvider, {
+  SignLocationState,
+} from "../components/SignTransaction/SignTransactionInfoProvider"
 
-export enum SignType {
-  Sign = "sign",
-  SignSwap = "sign-swap",
-  SignSpend = "sign-spend",
-  SignTransfer = "sign-transfer",
-}
-
-interface SignLocationState {
-  assetSymbol: string
-  amount: number
-  signType: SignType
-  to: string
-  value: string | number
-}
+export { SignType } from "../components/SignTransaction/SignTransactionInfoProvider"
 
 export default function SignTransaction({
   location,
@@ -46,19 +30,6 @@ export default function SignTransaction({
   const dispatch = useBackgroundDispatch()
   const transactionDetails = useBackgroundSelector(selectTransactionData)
 
-  const parsedTx = parseERC20Tx(transactionDetails?.input ?? "")
-  const isApproveTx = parsedTx?.name === "approve"
-
-  const getSignType = () => {
-    if (isApproveTx) {
-      return SignType.SignSpend
-    }
-    return SignType.Sign
-  }
-
-  const { assetSymbol, amount, to, value, signType } = location?.state ?? {
-    signType: getSignType(),
-  }
   const isTransactionDataReady = useBackgroundSelector(
     selectIsTransactionLoaded
   )
@@ -92,6 +63,7 @@ export default function SignTransaction({
         dispatch(broadcastSignedTransaction(signedTransaction))
       }
 
+      const assetSymbol = location.state?.assetSymbol
       // Request broadcast if not dApp...
       if (typeof assetSymbol !== "undefined") {
         history.push("/singleAsset", { symbol: assetSymbol })
@@ -100,14 +72,14 @@ export default function SignTransaction({
       }
     }
   }, [
-    isWaitingForKeyrings,
+    dispatch,
+    history,
     isTransactionSigned,
     isTransactionSigning,
-    history,
-    assetSymbol,
+    isWaitingForKeyrings,
+    location.state?.assetSymbol,
     shouldBroadcastOnSign,
     signedTransaction,
-    dispatch,
   ])
 
   if (isWaitingForKeyrings) {
@@ -137,78 +109,19 @@ export default function SignTransaction({
   const isWaitingForHardware =
     signerAccountTotal?.signingMethod?.type === "ledger" && isTransactionSigning
 
-  switch (signType) {
-    case SignType.SignSwap:
-      return (
+  return (
+    <SignTransactionInfoProvider location={location}>
+      {({ title, infoBlock, textualInfoBlock, confirmButtonLabel }) => (
         <SignTransactionContainer
           signerAccountTotal={signerAccountTotal}
-          title="Swap assets"
+          title={title}
           isWaitingForHardware={isWaitingForHardware}
-          infoBlock={<SignTransactionSwapAssetBlock />}
-          confirmButtonLabel="Confirm"
+          infoBlock={isWaitingForHardware ? textualInfoBlock : infoBlock}
+          confirmButtonLabel={confirmButtonLabel}
           handleConfirm={handleConfirm}
           handleReject={handleReject}
         />
-      )
-    case SignType.SignSpend:
-      return (
-        <SignTransactionContainer
-          signerAccountTotal={signerAccountTotal}
-          title="Approve asset spend"
-          isWaitingForHardware={isWaitingForHardware}
-          infoBlock={
-            <SignTransactionApproveSpendAssetBlock
-              transactionDetails={transactionDetails}
-              parsedTx={parsedTx}
-            />
-          }
-          confirmButtonLabel="Approve"
-          handleConfirm={handleConfirm}
-          handleReject={handleReject}
-        />
-      )
-    case SignType.SignTransfer:
-      return (
-        <SignTransactionContainer
-          signerAccountTotal={signerAccountTotal}
-          title="Sign Transfer"
-          isWaitingForHardware={isWaitingForHardware}
-          infoBlock={
-            isWaitingForHardware ? (
-              <SignTransactionTransferDetail
-                token={assetSymbol ?? ""}
-                amount={amount ?? 0}
-                destination={to ?? ""}
-              />
-            ) : (
-              <SignTransactionTransferBlock
-                token={assetSymbol ?? ""}
-                amount={amount ?? 0}
-                destination={to ?? ""}
-                localizedValue={value ?? ""}
-              />
-            )
-          }
-          confirmButtonLabel="Sign"
-          handleConfirm={handleConfirm}
-          handleReject={handleReject}
-        />
-      )
-    case SignType.Sign:
-      return (
-        <SignTransactionContainer
-          signerAccountTotal={signerAccountTotal}
-          title="Sign Transaction"
-          isWaitingForHardware={isWaitingForHardware}
-          infoBlock={
-            <SignTransactionSignBlock transactionDetails={transactionDetails} />
-          }
-          confirmButtonLabel="Sign"
-          handleConfirm={handleConfirm}
-          handleReject={handleReject}
-        />
-      )
-    default:
-      return <></>
-  }
+      )}
+    </SignTransactionInfoProvider>
+  )
 }
